requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
django-filter==23.3
django-environ==0.11.2
psycopg2-binary==2.9.9
Pillow==10.1.0
gunicorn==21.2.0
whitenoise==6.5.0
python-dotenv==1.0.0
django-storages==1.14.2
boto3==1.29.3
pytest==7.4.3
pytest-django==4.7.0
coverage==7.3.2
black==23.11.0
flake8==6.1.0
isort==5.12.0 